Overhaul Go To Definition
